requirements.txt
streamlit>=1.28.0
openai-agents>=0.0.6
openai>=1.68.0
python-dotenv>=1.0.0
